requirements.txt
# Python requirements for p(Doom)1 Website

# ============================================================================
# API Server v1 (Bridge) - No dependencies
# ============================================================================
# The bridge API server (scripts/api-server.py) uses only stdlib:
# - http.server for HTTP handling
# - json for data parsing
# - pathlib for file operations
# - datetime for timestamps

# ============================================================================
# API Server v2 (Production) - PostgreSQL + JWT
# ============================================================================

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Authentication
PyJWT==2.8.0           # JSON Web Tokens

# Environment configuration
python-dotenv==1.0.0   # Load .env files

# ============================================================================
# Optional Dependencies (Future Features)
# ============================================================================

# redis==5.0.1         # Redis caching (if needed later)
# requests==2.31.0     # HTTP client for external APIs
# gunicorn==21.2.0     # Production WSGI server (alternative to stdlib http.server)
